requirements.txt
flask==2.0.1
werkzeug==2.0.1
PyJWT==1.7.1
boto3==1.17.0
requests==2.25.1
